Text Anslyticd Project
This is a project for Text Analytics ISM 6930. We have created a dashboard using dash, plotly and deployed it on heroku cpud service . This Dashboard contains the change in sentiment of web streaming services over time and the features that help us identify the churners which help us predict the likelihood of churn. We choose to focus on the top 4 web streaming services (Amazon Prime Video, Hulu, Disney Plus, & Netflix).